Changed byte store writing strategy.
Large change of how data is written to file. Old approach caused
very large file formation when used with contiguous SChunks, as every
partially written chunk probably stayed where it was put initially.
Chunks are now cached in memory until full and then written to the
SChunk. There should be a performance increase, but it is not as
clear.